chore: break out implementation modules
