+++ src/convert/dummy.rs
@@ -1,0 +1,12 @@
+use anyhow::Result;
+use json::JsonValue;
+
+use super::*;
+
+pub struct DummyConverter;
+
+impl JsonConverter for DummyConverter {
+    fn convert_json(&self, input: JsonValue) -> Result<JsonValue> {
+        Ok(input)
+    }
+}
